Rename utils, update hook, and use readable name
Move matchCardUtils from Components/ to utils/ and update MatchCard import accordingly. Refactor useTournamentMatches: rename parameter to `competition`, rename loading state to `isLoading`, update getPredictions call and effect dependency. Replace hardcoded hero name with a human-readable `tournamentName` derived from the competition id (replaces underscores with spaces). These changes tidy up file organization, improve naming consistency, and surface a readable tournament title in the UI.
