+++ frontend/src/pages/TournamentPage/hooks/useTournamentMatches.ts
@@ -2,25 +2,25 @@ import { useEffect, useState } from "react";
 import { getPredictions } from "../../../services/predictionsApi";
 import type { PredictionPageMatch } from "../../../services/predictionsApi";
 
-export function useTournamentMatches(competitionId: string) {
+export function useTournamentMatches(competition: string) {
     const [matches, setMatches] = useState<PredictionPageMatch[]>([]);
-    const [isLoading, setLoading] = useState(true);
+    const [isLoading, setIsLoading] = useState(true);
     const [error, setError] = useState<string | null>(null);
 
     useEffect(() => {
         async function fetchMatches() {
             try {
-                const predictions = await getPredictions(competitionId);
+                const predictions = await getPredictions(competition);
                 setMatches(predictions.matches);
             } catch (err) {
                 setError("Failed to fetch matches");
             } finally {
-                setLoading(false);
+                setIsLoading(false);
             }
         }
 
         fetchMatches();
-    }, [competitionId]);
+    }, [competition]);
 
     return { matches, isLoading, error };
 }
